DNM configuration changes for QSPI storage of secondary image
This must be used with the corresponding branch of
https://github.com/pabigot/zephyr and works on nrf52840dk_nrf52840.

Significant configuration changes are made in the zephyr board
devicetree directly rather than using overlay devicetrees and
configurations because the mcuboot build process provides its own dts
overlay which causes any other to be ignored, and mcuboot and mcumgr
must agree on the layout.

Depends-on: https://github.com/zephyrproject-rtos/zephyr/pull/25971

## MCUBOOT

Configure to use the QSPI driver for flash, which requires thread
support and custom page size.  mcuboot will write BOOT_IMG_MAGIC
records that are stored in flash, so it needs a stack buffer for the
transfer.

## Zephyr

### board devicetree

Updated to use a secondary UART for console, so we can see messages
while still using serial mcumgr on the primary UART.

Use of QSPI overruns the 48 KiBy existing mcuboot partition, so it's
expanded to 64 KiBy.  Slot sizes are slightly adjusted just to make
things line up; may not be necessary.

The original slot1_partition on NVMC is commented out, and a
replacement added to the MX25R64 node.

### samples/mgmt/mcumgr

SMP_SHELL is disabled because it appeared to cause problems; see issue
#25294.

Kconfigs are added to match the QSPI functionality for mcuboot.
mcumgr also writes from a NVMC-stored block.

### hello_world

This is converted to an MCUBOOT-compatible to provide a small
application for upgrade testing.